python-backend/requirements.txt
# Excel 处理核心库
openpyxl==3.1.2
xlrd==2.0.1
xlwt==1.3.0
pandas==2.2.0
XlsxWriter==3.1.9

# 图像处理
Pillow==10.2.0

# Windows COM 组件（仅 Windows）
pywin32==306; sys_platform == 'win32'

# 测试框架
pytest==8.0.0
hypothesis==6.98.0
pytest-benchmark==4.0.0

# 打包工具
pyinstaller==6.3.0
